feat: introduce FieldsWithDefaultValue type
