+++ packages/client/src/v2-events/utils.ts
@@ -13,14 +13,6 @@ import { v4 as uuid } from 'uuid'
 import {
   ActionDocument,
   EventIndex,
-  FieldValue,
-  FieldType,
-  FieldConfigDefaultValue,
-  isTemplateVariable,
-  mapFieldTypeToZod,
-  isFieldValueWithoutTemplates,
-  compositeFieldTypes,
-  SystemVariables,
   Scope,
   ActionScopes,
   WorkqueueConfigWithoutQuery,
@@ -29,9 +21,6 @@ import {
   SystemRole,
   Location,
   UserOrSystem,
-  InteractiveFieldType,
-  FieldConfig,
-  TextField,
   flattenEntries,
   EventMetadataDateFieldId
 } from '@opencrvs/commons/client'
@@ -122,96 +111,6 @@ export function isTemporaryId(id: string) {
 
 export function createTemporaryId() {
   return `tmp-${uuid()}` as UUID
-}
-
-function isTextField(field: FieldConfig): field is TextField {
-  return field.type === FieldType.TEXT
-}
-
-/**
- *
- * @param fieldType: The type of the field.
- * @param currentValue: The current value of the field.
- * @param defaultValue: Configured default value from the country configuration.
- * @param meta: Metadata fields such as '$user', '$event', and others.
- *
- * @returns Resolves template variables in the default value and returns the resolved value.
- */
-export function replacePlaceholders({
-  field,
-  currentValue,
-  defaultValue,
-  systemVariables
-}: {
-  field: InteractiveFieldType
-  currentValue?: FieldValue
-  defaultValue?: FieldConfigDefaultValue
-  systemVariables: SystemVariables
-}): FieldValue | undefined {
-  if (currentValue) {
-    return currentValue
-  }
-
-  if (!defaultValue) {
-    return undefined
-  }
-
-  if (isFieldValueWithoutTemplates(defaultValue)) {
-    return defaultValue
-  }
-
-  if (isTemplateVariable(defaultValue)) {
-    const resolvedValue = get(systemVariables, defaultValue)
-    const validator = mapFieldTypeToZod(field)
-
-    const parsedValue = validator.safeParse(resolvedValue)
-
-    if (parsedValue.success) {
-      return parsedValue.data as FieldValue
-    }
-
-    throw new Error(`Could not resolve ${defaultValue}: ${parsedValue.error}`)
-  }
-
-  if (
-    compositeFieldTypes.some((ft) => ft === field.type) &&
-    typeof defaultValue === 'object'
-  ) {
-    /**
-     * defaultValue is typically an ADDRESS, FILE, or FILE_WITH_OPTIONS.
-     * Some STRING values within the defaultValue object may contain template variables (prefixed with $).
-     */
-    const result = { ...defaultValue }
-
-    // @TODO: This resolves template variables in the first level of the object. In the future, we might need to extend it to arbitrary depth.
-    for (const [key, val] of Object.entries(result)) {
-      if (val && isTemplateVariable(val) && isTextField(field)) {
-        const resolvedValue = get(systemVariables, val)
-        // For now, we only support resolving template variables for text fields.
-        const validator = mapFieldTypeToZod(field)
-        const parsedValue = validator.safeParse(resolvedValue)
-        if (parsedValue.success && parsedValue.data) {
-          result[key] = resolvedValue
-        } else {
-          throw new Error(`Could not resolve ${key}: ${parsedValue.error}`)
-        }
-      }
-    }
-
-    const resultValidator = mapFieldTypeToZod(field)
-    const parsedResult = resultValidator.safeParse(result)
-    if (parsedResult.success) {
-      return result as FieldValue
-    }
-    throw new Error(
-      `Could not resolve ${field.type}: ${JSON.stringify(
-        defaultValue
-      )}. Error: ${parsedResult.error}`
-    )
-  }
-  throw new Error(
-    `Could not resolve ${field.type}: ${JSON.stringify(defaultValue)}`
-  )
 }
 
 export const AssignmentStatus = {
